+++ src/app/components/logo-main/logo-main.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, ViewEncapsulation } from '@angular/core';
+import { Component, Input, OnInit, ViewEncapsulation } from '@angular/core';
 
 @Component({
     selector: 'app-logo-main',
@@ -9,6 +9,9 @@ import { Component, OnInit, ViewEncapsulation } from '@angular/core';
 })
 export class LogoMainComponent implements OnInit {
     constructor() {}
+
+    @Input()
+    public showBackArrow: boolean;
 
     public ngOnInit(): void {}
 }